requirements.txt
seaborn
statsmodels
scikit-learn
numpy
pandas
matplotlib
pillow
requests

